requirements.txt
altgraph==0.17.4
bidict==0.23.1
certifi==2024.8.30
charset-normalizer==3.4.0
h11==0.14.0
idna==3.10
macholib==1.16.3
packaging==24.2
psutil==6.1.0
pyinstaller==6.11.1
pyinstaller-hooks-contrib==2024.10
python-engineio==4.10.1
python-socketio==5.11.4
requests==2.32.3
setuptools==75.3.0
simple-websocket==1.1.0
urllib3==2.2.3
websocket-client==1.8.0
wsproto==1.2.0
